Add Authorisation to Booking routes

+++ routes/bookingsRoutes.js
@@ -1,16 +1,34 @@
+// Imprt Files
 import { Router } from 'express'
 import db from '../db.js'
+import jwt from 'jsonwebtoken'
+import { jwtSecret } from '../secrets.js'
+
+// Start app
 const router = Router()
+
+// USER VERIFICATION
+function decodeToken(cookie) {
+  const decodedToken = jwt.verify(cookie, jwtSecret)
+  if (!cookie || !decodedToken || !decodedToken.user_id) {
+    throw new Error('Invalid authentication token')
+  }
+  return decodedToken
+}
 
 // POST API Calls
 // Add new booking to database
 router.post('/bookings', async (req, res) => {
   try {
-    // INSERT Query with post parameters
+    // USER VERIFICATION
+    const decodedToken = decodeToken(req.cookies.jwt)
+
+    // CREATE NEW BOOKING
+    // INSERT Query with post parameters and forcing guest_id to be the user making the post request
     const { rows } = await db.query(`
       INSERT INTO bookings (guest_id, house_id, check_in_date, check_out_date, total_price)
       VALUES (
-        ${req.body.guest_id},
+        ${decodedToken.user_id},
         ${req.body.house_id},
         '${req.body.check_in_date}',
         '${req.body.check_out_date}',
@@ -28,13 +46,13 @@ router.post('/bookings', async (req, res) => {
 // Fetch bookings from database
 router.get('/bookings', async (req, res) => {
   try {
-    // Select Query based on whether there is a userID or not
-    let finalQuery = !req.query.userid
-      ? 'SELECT * FROM bookings ORDER by check_in_date DESC'
-      : `SELECT * FROM bookings WHERE guest_id = ${req.query.userid} ORDER BY check_in_date DESC`
+    // USER VERIFICATION
+    const decodedToken = decodeToken(req.cookies.jwt)
 
-    // Deconstruct rows using finalQuery
-    const { rows } = await db.query(finalQuery)
+    // Fetch bookings from DB based on user_id making the request
+    const { rows } = await db.query(
+      `SELECT * FROM bookings WHERE guest_id = ${decodedToken.user_id} ORDER BY check_in_date DESC`
+    )
 
     // Throw Error if the userid provided is not on database or
     if (!rows.length) {
@@ -46,13 +64,23 @@ router.get('/bookings', async (req, res) => {
   }
 })
 
-// Fetch bookings with index of 1
+// Fetch bookings for logged in User
 router.get('/bookings/:bookingId', async (req, res) => {
   const { bookingId } = req.params
   try {
+    // USER VERIFICATION
+    const decodedToken = decodeToken(req.cookies.jwt)
+
+    // Fetch Booking
     const { rows } = await db.query(
       `SELECT * FROM bookings WHERE booking_id = ${bookingId}`
     )
+    // Verify user is same as booking requested
+    if (rows[0].guest_id !== decodedToken.user_id) {
+      throw new Error('You are not authorized')
+    }
+
+    // Verify a booking was found
     if (!rows.length) {
       throw new Error('Booking Id not found.')
     }
@@ -64,19 +92,36 @@ router.get('/bookings/:bookingId', async (req, res) => {
 
 // Delete booking
 router.delete('/bookings/:bookingId', async (req, res) => {
+  const { bookingId } = req.params
   try {
-    const { rows } = await db.query(`
-  DELETE FROM bookings WHERE booking_id = ${req.params.bookingId}
-  RETURNING *`)
-    if (!rows.length) {
-      throw new Error('Booking not found')
+    // USER VERIFICATION
+    const decodedToken = decodeToken(req.cookies.jwt)
+
+    // Fetch booking from DB
+    const { rows } = await db.query(
+      `SELECT guest_id FROM bookings WHERE booking_id = ${bookingId}`
+    )
+    // Verify loggedInUser is the one requesting deletion
+    if (rows[0].guest_id === decodedToken.user_id) {
+      // Delete query
+      const { rows } = await db.query(`
+        DELETE FROM bookings WHERE booking_id = ${req.params.bookingId}
+        RETURNING *`)
+
+      // Verify there is a booking on that booking_id
+      if (!rows.length) {
+        throw new Error('Booking not found')
+      }
+
+      // Send response
+      res.json('Booking successfully deleted')
+    } else {
+      // Throw error when guest_id on booking is not the same as user requesting the deletion
+      throw new Error('You are not authorized')
     }
-    res.json('Booking successfully deleted')
   } catch (err) {
-    console.log(err.message)
     res.json(err.message)
   }
 })
-
 
 export default router
